requirements.txt
# Core dependencies for video processing
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
Pillow==10.0.1
numpy==1.24.3

# AI/ML dependencies
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.4
transformers==4.33.2
accelerate==0.23.0

# Web framework for API
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6

# Utilities
tqdm==4.66.1
click==8.1.7
python-dotenv==1.0.0
psutil==5.9.5

# Development dependencies
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0